test: split command-simplified + isolate upload (mock pollution fix)
Two changes that together get the main suite from 4 fail → 0 fail:

1. Move test/upload.test.ts → test/isolated/upload.test.ts
   Root cause: line 24 mocked the entire `os` module process-globally
   (`mock.module("os", () => ({ homedir: () => TEST_HOME }))`). Bun's
   mock.module is process-wide, so any subsequent test calling
   `homedir()` got the upload tmpdir instead of the real $HOME. This
   broke every channelEnv tilde-expansion test (#1135) + the new
   permissionMode-with-channelEnv test (#1146).

   Move to isolated/ runs each test file in its own subprocess →
   pollution can't leak across files.

2. Split command-simplified.test.ts (342 LOC, 28 tests) into 3 files
   per the small-modular-tests pattern:
   - test/build-command-contract.test.ts (12 tests, post-#541)
   - test/build-command-channel-env.test.ts (10 tests, #1135 + #1148)
   - test/build-command-permission-mode.test.ts (6 tests, #1146)

   Each file has its own minimal mock setup. Smaller files = tighter
   mock surface = better resilience to future cross-file pollution.

Verified: isolated 79/79, main 1217 pass / 0 fail / 12 skip.
